Remove alt text from images in index.html
The following images have textual alternatives. Please review that the textual alternatives make sense considering the purpose of the image in the context of the page and that phrases like “Image of …” are avoided.

Note that iconic images that are redundant with text next to them or purely decorative should have alt="" instead.